+++ src/popt/optimization_methods/linesearch.py
@@ -231,7 +231,7 @@ class LineSearch(OptimizerBase):
             callback=callback,
             **options
         )
-        optimizer.optimization_loop()
+        optimizer.run_optimization()
         return optimizer.optimize_results