+++ src/popt/optimization_methods/trust_region.py
@@ -228,7 +228,7 @@ class TrustRegion(OptimizerBase):
             callback=callback,
             **options,
         )
-        optimizer.optimization_loop()
+        optimizer.run_optimization()
         return optimizer.optimize_results
 
     def update_step(self) -> bool: